refactor: api 명세서 형식으로 타입 재정립
