+++ src/app/wines/components/SuggestedWine/type.ts
@@ -1,0 +1,6 @@
+import { WineResponse } from "@/app/wines/type";
+
+export type SuggestedWineProps = Pick<
+  WineResponse,
+  'id' | 'name' | 'region' | 'image'
+>;
