--- src/app/wines/components/SuggestedWineList/type.ts
@@ -1,5 +1,0 @@
-import { SuggestedWineProps } from '@/app/wines/type';
-
-export interface SuggestedWineListProps {
-  wines: SuggestedWineProps[];
-}
