+++ src/app/wines/components/WineCard/type.ts
@@ -1,0 +1,3 @@
+import { WineResponse } from "@/app/wines/type";
+
+export type WineCardProps = WineResponse;
