+++ src/app/wines/type.ts
@@ -1,29 +1,36 @@
-import { WineType } from '@/constants/chips';
+import type { WineType } from '@/constants/chips';
 
-export type SuggestedWineProps = {
+export interface User {
+  id: number;
+  nickname: string;
+  image: string | null;
+}
+
+export interface RecentReview {
+  user: User;
+  updatedAt: string;
+  createdAt: string;
+  content: string;
+  aroma: string[];
+  rating: number;
+  id: number;
+}
+
+export interface WineResponse {
   id: number;
   name: string;
   region: string;
-  image: string;
-};
-
-export type WineCardProps = SuggestedWineProps & {
+  image: string | null;
   price: number;
   type: WineType;
   avgRating: number;
   reviewCount: number;
-  recentReview: {
-    user: {
-      id: number;
-      nickname: string;
-      image: string;
-    };
-    updatedAt: string;
-    createdAt: string;
-    content: string;
-    aroma: string[];
-    rating: number;
-    id: number;
-  };
+  recentReview: RecentReview;
   userId: number;
-};
+}
+
+export interface WinesResponse {
+  totalCount: number;
+  nextCursor: number | null;
+  list: WineResponse[];
+}
